Move print_progress to console.py & get_seconds to ffmpeg.py
This fixes the cyclic-import R0401 (purewebm.ffmpeg -> purewebm.purewebm
-> purewebm.webm)
